requirements.txt
fastapi==0.97.0
uvicorn==0.21.0
tortoise-orm==0.19.3
python-dotenv==1.0.0